requirements.txt
pytorch-argus==1.0.0
numpy==1.24.2
scipy==1.10.1
hydra-core===1.3.2
opencv-python==4.7.0.72
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
notebook==6.5.3
tqdm==4.66.4
optuna<3.0.0
SoccerNet==0.1.51
lsq-ellipse==2.2.1
hydra-optuna-sweeper==1.2.0
torchvision